requirements.txt
fastapi
uvicorn>=0.20.0
aiofiles>=3.8.1
python-multipart>=0.0.5
jinja2>=3.1.2
requests
python-dotenv
quantlib-python
pandas
numpy
pandas_market_calendars
thetadata>=0.8.9
yfinance>=0.2.12
polygon-api-client
scipy
matplotlib
supabase